Observer Pattern Chain of Responsibility Facade Pattern Factory Pattern Visitor Pattern
Done
-------

All patterns are working, but the Employee App is not finished.
Still need to validate user input and display earnings on the table.

+++ src/factory_pattern/visitor_pattern/EmployeeVisitorImpl.java
@@ -6,7 +6,7 @@ public class EmployeeVisitorImpl implements EmployeeVisitor
     @Override
     public double visit(HourlyEmployee hourlyEmployee)
     {
-        int hours = hourlyEmployee.getHoursWorked();
+        double hours = hourlyEmployee.getHoursWorked();
         double salary = hourlyEmployee.getSalary();
 
         if (hours <= 40)
